AeroLeads/01_LinkedIn_Scrapping/requirements.txt
streamlit==1.31.0
selenium==4.16.0
webdriver-manager==4.0.1
pandas==2.1.4
python-dotenv==1.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
